Protect API Routes (#109)
* Create tests for organizations backend

* Add authentication security for event api routes

* Add security to org role api/services

This commit adds a check to ensure that users cannot create/delete org
roles without the proper permissions in the backend. This ensures that
people cannot go to the docs page and create/delete org roles.

* Protect organization api routes

* Fix Org Editor Permissions

This commit fixes the org editor permissions. Previously, the "create"
version of the page would be visible to anyone who is an organization
manager. This commit ensures that it is only available to admins
specifically.

* Add security to mutating registration API routes

This commit adds security/permission checks to the mutating registration
API routes (create/update status/delete/clear all) in order to prevent
users from going into the docs and making changes to the registrations
table.

* Fix whitespace

* Fix tags and add comment

This commit fixes the tags in the api route files so they do not include
"Detail" as a suffix in order to improve the readability of the docs
page. It also adds a comment explaining the -1 org id in the org-editor
file, which is used to indicate that a new organization (which does not
yet have an id since it hasn't been added to the table) is being made.

* Fix Event API Tags

Edited Event tags from EventDetail back to Event.

---------

+++ backend/api/event.py
@@ -2,10 +2,19 @@ from fastapi import APIRouter, Depends, HTTPException
 from backend.services.event import EventService
 from backend.models.event import EventDetail, Event
 from datetime import datetime
+from fastapi.security import HTTPBearer
+from fastapi.security.http import HTTPAuthorizationCredentials
+import jwt
+from ..env import getenv
+
+from backend.services.user import UserService
 
 api = APIRouter(prefix="/api/events")
 
-@api.get("", tags=['EventDetail'])
+_JWT_SECRET = getenv('JWT_SECRET')
+_JST_ALGORITHM = 'HS256'
+
+@api.get("", tags=['Event'])
 def get_events(event_service: EventService = Depends()) -> list[EventDetail]:
     """
     Get all events
@@ -16,7 +25,7 @@ def get_events(event_service: EventService = Depends()) -> list[EventDetail]:
     # Return all events
     return event_service.all()
 
-@api.get("/org/{id}/", tags=['EventDetail'])
+@api.get("/org/{id}/", tags=['Event'])
 def get_events_org_id(org_id: int, event_service: EventService = Depends()) -> list[EventDetail]:
     """
     Get all events from an organization
@@ -27,7 +36,7 @@ def get_events_org_id(org_id: int, event_service: EventService = Depends()) -> l
     # Return all events
     return event_service.get_from_org_id(org_id)
 
-@api.get("/time/", tags=['EventDetail'])
+@api.get("/time/", tags=['Event'])
 def get_events_from_time_range(start: datetime, end: datetime, event_service: EventService = Depends()) -> list[EventDetail]:
     """
     Get all events from a time range
@@ -38,8 +47,10 @@ def get_events_from_time_range(start: datetime, end: datetime, event_service: Ev
     # Return all events
     return event_service.get_from_time_range(start, end)
 
-@api.post("", tags=['EventDetail'])
-def new_event(event: Event, event_service: EventService = Depends()) -> EventDetail:
+@api.post("", tags=['Event'])
+def new_event(event: Event, event_service: EventService = Depends(),
+              user_service: UserService = Depends(),
+              token: HTTPAuthorizationCredentials | None = Depends(HTTPBearer())) -> EventDetail:
     """
     Create event
     Returns:
@@ -47,15 +58,29 @@ def new_event(event: Event, event_service: EventService = Depends()) -> EventDet
     """
 
     # Try to create event
-    try:
-        # Return created event
-        return event_service.create(event)
-    except Exception as e:
-        # Raise 422 exception if creation fails
-        # - This would occur if the request body is shaped incorrectly
-        raise HTTPException(status_code=422, detail=str(e))
+    if token:
+        try:
+            auth_info = jwt.decode(
+                token.credentials, _JWT_SECRET, algorithms=[_JST_ALGORITHM]
+            )
+            user = user_service.get(auth_info['pid'])
+            if user:
+                org_roles = [org_role for org_role in user.organization_associations if
+                             org_role.org_id == event.org_id and org_role.membership_type > 0]
+                if(len(org_roles) <=0):
+                    raise HTTPException(status_code=401, detail="Unauthorized")
+                try:
+                    # Return created event
+                    return event_service.create(event)
+                except Exception as e:
+                    # Raise 422 exception if creation fails
+                    # - This would occur if the request body is shaped incorrectly
+                    raise HTTPException(status_code=422, detail=str(e))
+        except:
+            ...
+    raise HTTPException(status_code=401, detail='Unauthorized')
 
-@api.get("/{id}", responses={404: {"model": None}}, tags=['EventDetail'])
+@api.get("/{id}", responses={404: {"model": None}}, tags=['Event'])
 def get_event_from_id(id: int, event_service: EventService = Depends()) -> EventDetail:
     """
     Get event with matching id
@@ -72,34 +97,66 @@ def get_event_from_id(id: int, event_service: EventService = Depends()) -> Event
         # - This would occur if there is no response
         raise HTTPException(status_code=404, detail=str(e))
 
-@api.put("", responses={404: {"model": None}}, tags=['EventDetail'])
-def update_event(organization: EventDetail, event_service: EventService = Depends()) -> EventDetail:
+@api.put("", responses={404: {"model": None}}, tags=['Event'])
+def update_event(event: EventDetail, event_service: EventService = Depends(),
+              user_service: UserService = Depends(),
+              token: HTTPAuthorizationCredentials | None = Depends(HTTPBearer())) -> EventDetail:
     """
     Update event
     Returns:
         EventDetail: Updated event
     """
+    if token:
+        try:
+            auth_info = jwt.decode(
+                token.credentials, _JWT_SECRET, algorithms=[_JST_ALGORITHM]
+            )
+            user = user_service.get(auth_info['pid'])
+            if user:
+                org_roles = [org_role for org_role in user.organization_associations if
+                             org_role.org_id == event.org_id and org_role.membership_type > 0]
+                if(len(org_roles) <=0):
+                    raise HTTPException(status_code=401, detail="Unauthorized")
+                # Try to update event
+                try: 
+                    # Return updated event
+                    return event_service.update(event)
+                except Exception as e:
+                    # Raise 404 exception if search fails
+                    # - This would occur if there is no response
+                    raise HTTPException(status_code=404, detail=str(e))
+        except:
+            ...
+    raise HTTPException(status_code=401, detail='Unauthorized')
 
-    # Try to update event
-    try: 
-        # Return updated event
-        return event_service.update(organization)
-    except Exception as e:
-        # Raise 404 exception if search fails
-        # - This would occur if there is no response
-        raise HTTPException(status_code=404, detail=str(e))
-
-@api.delete("/{id}", tags=['EventDetail'])
-def delete_event(id: int, event_service = Depends(EventService)):
+@api.delete("/{id}", tags=['Event'])
+def delete_event(id: int, event_service = Depends(EventService),
+              user_service: UserService = Depends(),
+              token: HTTPAuthorizationCredentials | None = Depends(HTTPBearer())):
     """
     Delete event based on id
     """
-
-    # Try to delete event
-    try:
-        # Return deleted event
-        return event_service.delete(id)
-    except Exception as e:
-        # Raise 404 exception if search fails
-        # - This would occur if there is no response or if item to delete does not exist
-        raise HTTPException(status_code=404, detail=str(e))
+    if token:
+        try:
+            auth_info = jwt.decode(
+                token.credentials, _JWT_SECRET, algorithms=[_JST_ALGORITHM]
+            )
+            user = user_service.get(auth_info['pid'])
+            if user:
+                event = event_service.get_from_id(id)
+                org_roles = [org_role for org_role in user.organization_associations if
+                             org_role.org_id == event.org_id and org_role.membership_type > 0]
+                if(len(org_roles) <=0):
+                    raise HTTPException(status_code=401, detail="Unauthorized")
+                # Try to update event
+                # Try to delete event
+                try:
+                    # Return deleted event
+                    return event_service.delete(id)
+                except Exception as e:
+                    # Raise 404 exception if search fails
+                    # - This would occur if there is no response or if item to delete does not exist
+                    raise HTTPException(status_code=404, detail=str(e))
+        except:
+            ...
+    raise HTTPException(status_code=401, detail='Unauthorized')
